% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPlot.r
\name{pcaPlot}
\alias{pcaPlot}
\title{Principal Component Analysis (PCA) on Normalized RNA-seq Data}
\usage{
pcaPlot(dds, ntop = 500, title = "PCA Plot")
}
\arguments{
\item{dds}{: DESeqDataSet}

\item{ntop}{: Number of most variable genes to use for PCA (default is 500)}

\item{title}{: Optional title for plot}
}
\value{
a ggplot2 PCA plot
}
\description{
This function produces a PCA plot to explore sample relationships and
detect possible outliers of the DESeqDataSet file.
}
\examples{
library(DESeq2)
library(airway)
data(airway)
dds <- DESeqDataSet(airway, design = ~ dex)
pcaPlot(dds)

}
\references{
Love, M.I., Huber, W. & Anders, S. Moderated estimation of fold change and
dispersion for RNA-seq data with DESeq2. Genome Biol 15, 550 (2014).
https://doi.org/10.1186/s13059-014-0550-8

Morgan M, Obenchain V, Hester J, Pagès H (2025). SummarizedExperiment:
A container (S4 class) for matrix-like assays.
doi:10.18129/B9.bioc.SummarizedExperiment, R package version 1.40.0,
https://bioconductor.org/packages/SummarizedExperiment.

OpenAI. (2025). ChatGPT (GPT-5) https://chat.openai.com/

Wickham H (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. ISBN 978-3-319-24277-4, https://ggplot2.tidyverse.org.
}
