% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcPlot.r
\name{qcPlot}
\alias{qcPlot}
\title{Purpose: Quick Quality Control Plots for RNA-seq data before DESeq2 analysis
Author: Katie Lee
Date: November 4, 2025}
\usage{
qcPlot(dds, title = "Quality Control Summary", min_counts = 10)
}
\arguments{
\item{dds}{: DESeqDataSet}

\item{title}{: Optional title for plots}

\item{min_counts}{: Minimum total counts per gene to include (default = 10)}
}
\value{
two ggplot2 figures
}
\description{
This function produces a set of QC plots to examine the DESeqDataSet.
This includes the total library size per sample and distribution of
counts per sample in log.
}
\examples{
if (!requireNamespace("airway", quietly = TRUE)) {
  if (!requireNamespace("BiocManager", quietly = TRUE))
    install.packages("BiocManager")
  BiocManager::install("airway")
  }

library(DESeq2)
library(airway)
data(airway)
dds <- DESeqDataSet(airway, design = ~ dex)
DESeqPLUS::qcPlot(dds)

}
\references{
Love, M.I., Huber, W. & Anders, S. Moderated estimation of fold change and
dispersion for RNA-seq data with DESeq2. Genome Biol 15, 550 (2014).
https://doi.org/10.1186/s13059-014-0550-8

Wickham H (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. ISBN 978-3-319-24277-4, https://ggplot2.tidyverse.org.
}
