% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDESeqPLUS.r
\name{runDESeqPLUS}
\alias{runDESeqPLUS}
\title{Purpose: Shiny App for DESeqPLUS
Author: Katie Lee
Date: November 28, 2025}
\usage{
runDESeqPLUS()
}
\value{
launches Shiny app.
}
\description{
This function launches the Shiny app for DESEqPLUS.
}
\examples{

DESeqPLUS::runDESeqPLUS()

}
