% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.r
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Purpose: Volcano Plot from DESeq2 results
Author: Katie Lee
Date: November 4, 2025}
\usage{
volcanoPlot(
  res,
  lfc_threshold = 1,
  pval_threshold = 0.05,
  title = "Volcano Plot"
)
}
\arguments{
\item{res}{: DESeqResults}

\item{lfc_threshold}{: Log2 fold change threshold for highlighting (default is 0)}

\item{pval_threshold}{: Adjusted p-value threshold (default is 0.05)}

\item{title}{: Optional title for the plot}
}
\value{
a ggplot2 volcano plot
}
\description{
This function produces a volcano plot showing log2 fold changes vs -log10 p-values.
Points with p-value less than threshold are colored in red, and others are grey.
}
\examples{
library(DESeq2)
library(airway)
data(airway)
dds <- DESeqDataSet(airway, design = ~ dex)
dds <- DESeq(dds)
res <- results(dds)
DESeqPLUS::volcanoPlot(res)

}
\references{
Wickham H (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. ISBN 978-3-319-24277-4, https://ggplot2.tidyverse.org.
}
