% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDESeqPLUS.r
\name{runDESeqPLUS}
\alias{runDESeqPLUS}
\title{Shiny App for DESeqPLUS}
\usage{
runDESeqPLUS()
}
\value{
launches Shiny app.
}
\description{
This function launches the Shiny app for DESEqPLUS.
}
\examples{

DESeqPLUS::runDESeqPLUS()

}
